+++ packages/antd-management-fast-component/src/components/FieldExtra/SelectFieldExtra/BaseSelectFieldExtra/index.jsx
@@ -9,11 +9,7 @@ import {
   isFunction,
   logTrace,
   mergeArrowText,
-  showSimpleRuntimeError,
-  toNumber,
 } from 'easy-soft-utility';
-
-import { selectModeCollection } from 'antd-management-fast-common';
 
 import { BaseComponent } from '../../../../bases/BaseComponent';
 import { FormExtra } from '../../../FormExtra';
@@ -44,54 +40,20 @@ class BaseSelectFieldExtra extends BaseComponent {
   };
 
   showSelect = () => {
-    const { selectMode } = this.props;
-
-    if (toNumber(selectMode) === selectModeCollection.drawer) {
-      this.showDrawer();
-    }
-
-    if (toNumber(selectMode) === selectModeCollection.modal) {
-      this.showModal();
-    }
+    this.openSelector();
   };
 
-  openDrawer = () => {
-    throw new Error(this.buildOverloadErrorText('openDrawer'));
+  openSelector = () => {
+    throw new Error(this.buildOverloadErrorText('openSelector'));
   };
 
-  showDrawer = () => {
-    this.openDrawer();
-  };
-
-  afterDrawerSelectSuccess = (o) => {
+  afterSelectSuccess = (o) => {
     logTrace(
       mergeArrowText(
         this.componentName,
         primaryCallName,
         'trigger',
-        'afterDrawerSelectSuccess',
-        'afterSelectSuccessCore',
-      ),
-    );
-
-    this.afterSelectSuccessCore(o);
-  };
-
-  openModal = () => {
-    throw new Error(this.buildOverloadErrorText('openModal'));
-  };
-
-  showModal = () => {
-    this.openModal();
-  };
-
-  afterModalSelectSuccess = (o) => {
-    logTrace(
-      mergeArrowText(
-        this.componentName,
-        primaryCallName,
-        'trigger',
-        'afterModalSelectSuccess',
+        'afterSelectSuccess',
         'afterSelectSuccessCore',
       ),
     );
@@ -122,12 +84,8 @@ class BaseSelectFieldExtra extends BaseComponent {
     throw new Error(this.buildOverloadErrorText('selectValueText'));
   };
 
-  renderPresetSelectDrawer = () => {
-    throw new Error(this.buildOverloadErrorText('renderPresetSelectDrawer'));
-  };
-
-  renderPresetSelectModal = () => {
-    throw new Error(this.buildOverloadErrorText('renderPresetSelectModal'));
+  renderPresetSelector = () => {
+    throw new Error(this.buildOverloadErrorText('renderPresetSelector'));
   };
 
   renderPresetField = () => {
@@ -207,21 +165,7 @@ class BaseSelectFieldExtra extends BaseComponent {
   };
 
   renderPresetOther = () => {
-    const { selectMode } = this.props;
-
-    if (toNumber(selectMode) === selectModeCollection.drawer) {
-      return this.renderPresetSelectDrawer();
-    }
-
-    if (toNumber(selectMode) === selectModeCollection.modal) {
-      return this.renderPresetSelectModal();
-    }
-
-    const text = '无效的选择项渲染模式';
-
-    showSimpleRuntimeError(text);
-
-    return null;
+    return this.renderPresetSelector();
   };
 
   renderFurther() {
@@ -243,7 +187,6 @@ BaseSelectFieldExtra.defaultProps = {
   formItemLayout: null,
   required: false,
   showClear: true,
-  selectMode: selectModeCollection.drawer,
 };
 
 export { BaseSelectFieldExtra };