+++ packages/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.8.50
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.37
+
 ## 1.8.49
 
 ### Patch Changes