+++ packages/example-remote/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.4.48
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.37
+
 ## 1.4.47
 
 ### Patch Changes